Fix: do not use charmhelpers in non-charm context
In change I60141397f39e3b1b0274230db8d984934c98a08d charmhelper
library is being used in the rabbitmq queue nrpe check. This is
problematic as the check does not actually run in a charm context and
therefore does not have access to the charm environment such as the
current config. Additionally an issue in collating check results had
been introduced.

This change aims to fix these issues. Instead of using the charmhelper
library, the cronspec is read out from the cron job definition
itself, and the series is probed from /etc/lsb-release

Change-Id: I952aeda31e997ccadb6cff62e3b0d46349650979