Install python3-croniter for nrpe checks
Change ab79c3ee introduced the croniter library for interacting
with cron. However, the dependent package was not included in
a list of installation packages nor installed on the upgrade
path. This patch adds the python3-croniter package for upgrade
and installation paths to allow the nrpe checks to continue
to function.

Change-Id: I363f8e04a3850eecf219c5a26ab9137f217a0f36
Closes-Bug: #1928802