Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: Idd4eb37dd413b49a99b8b81fd15d0b45f75ae97e
(cherry picked from commit 84b8605ed6cfd439a9163b5e41eee24a4bcc2eb6)