Don't overwrite last relation key when username not found
When rotating a password, the code updates the password on the relation
bag for the associated relation. However, if the username wasn't found
in the relation data (e.g. if it was in app-data instead) then the code
unfortunately overwrote the last key it looked at (and this was,
randomly, private-address). This was due to a bug in the code. This
patch fixes that problem.

However, the charm (or at least certainly the rotating passwords code)
doesn't support app data bags as it doesn't find the matching username
to update the relation data. This means that it doesn't support rotating
passwords with charms that use app-data. This is the note added to the
README.

Change-Id: I05dac3ae89318ceb28724f4a75d1377a62d32d1c
Closes-Bug: #2051365