Add consumer-timeout config option
Exposes the consumer timeout configuration value that was changed in the jammy
version of rabbitmq. The value used to default to unlimited, but was changed to
30 minutes. This causes clients to timeout for longer-running jobs.

This change doesn't alter the payload's default value, it override it only when
the user provides a value via juju config.

Closes-Bug: #2020909
Related-Bug: #2067424

Change-Id: I552c04c964c98b2a700cd47e98a2ae491f5fd47b
(cherry picked from commit edf7791f4032951ccfaa336c7ccd734bdef1873d)