Allow for long rabbitmq node names
In some environments, the call to socket.gethostname resolves to the
fqdn. This causes the charm to use the fqdn for the rabbit hostname,
but rabbit always uses the equivalent of `hostname -s` by default. The
charm already uses the results of socket.gethostname() in relation data
everywhere, and some users have applied a workaround of
RABBITMQ_USE_LONGNAME to the rabbitmq-env.conf file in order to allow
the nodes to start and the charm to manage it.

This change detects the scenario where the socket.gethostname() call
results in a name that contains a '.' and treats this as the long node
name scenario. It will detect this on the install hook and render the
necessary configuration files and restart the rabbit service to ensure
that the node specified by the fqdn is used. This allows for the charm
to manage the rabbit cluster in these particular scenarios.

What this change does not deal with, is the potential that the hostname
resolution can change from hostname to fqdn (or vice versa). This would
require that the node be gracefully renamed in the rabbitmq-cluster and
is a much bigger/more involved change. As the current charm makes
assumptions that the hostname is stable, this is left to future
enhancements for the charm.

Related-Bug: #1976523
Change-Id: I27013e0f998fd0f0edf31b029a857af652ec6e8e